changing project to use C exclusively instead of C++
